# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ops & Planning API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  stream_overseer: "stream_overseer",
  leader: "leader",
} as const;

export interface AuthStatus {
  needsSetup: boolean;
}

export interface User {
  id: string;
  email: string;
  name: string;
  initials: string;
  department: string;
  role: UserRole;
  streamId?: string | null;
  teamId?: string | null;
  active: boolean;
  invitedByName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AuthSession {
  token: string;
  user: User;
}

export interface LoginBody {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface SetupBody {
  email: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 8 */
  password: string;
  department?: string;
  /** @minLength 1 */
  setupSecret: string;
}

export interface AcceptInviteBody {
  /** @minLength 1 */
  token: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 8 */
  password: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  /** @minLength 1 */
  token: string;
  /** @minLength 8 */
  password: string;
}

export interface CreateInviteBody {
  email: string;
  /** @minLength 1 */
  name: string;
  role: UserRole;
  department?: string;
  streamId?: string | null;
  teamId?: string | null;
}

export interface CreateInviteResponse {
  message: string;
  acceptUrl: string;
}

export interface InvitePreview {
  email: string;
  name: string;
  role: UserRole;
  department: string;
}

export interface CreateUserBody {
  email: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 6 */
  password: string;
  role: UserRole;
  department?: string;
  streamId?: string | null;
  teamId?: string | null;
}

export interface UpdateUserRoleBody {
  role: UserRole;
  streamId?: string | null;
  teamId?: string | null;
}

export interface Stream {
  id: string;
  name: string;
  description: string;
  programmeId: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStreamBody {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  programmeId: string;
}

export interface UpdateStreamBody {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export interface Team {
  id: string;
  name: string;
  streamId?: string | null;
  leaderId?: string | null;
  functionLabel?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type TeamWithStreamName = Team & {
  streamName?: string | null;
};

export interface CreateTeamBody {
  /** @minLength 1 */
  name: string;
  streamId?: string | null;
  leaderId?: string | null;
  functionLabel?: string;
}

export interface UpdateTeamBody {
  /** @minLength 1 */
  name?: string;
  streamId?: string | null;
  leaderId?: string | null;
  functionLabel?: string;
}

export interface AssignLeaderBody {
  leaderId?: string | null;
}

export interface Member {
  id: string;
  name: string;
  roleLabel?: string | null;
  teamId: string;
  createdAt: string;
}

export interface CreateMemberBody {
  /** @minLength 1 */
  name: string;
  roleLabel?: string;
}

export interface UpdateMemberBody {
  /** @minLength 1 */
  name?: string;
  roleLabel?: string;
  teamId?: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  not_started: "not_started",
  in_progress: "in_progress",
  at_risk: "at_risk",
  completed: "completed",
} as const;

export interface Project {
  id: string;
  teamId: string;
  title: string;
  description: string;
  status: ProjectStatus;
  color: string;
  phase: string;
  dueDate?: string | null;
  notes: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export type ProjectWithTeamName = Project & {
  teamName?: string | null;
};

export interface CreateProjectBody {
  /** @minLength 1 */
  teamId: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: ProjectStatus;
  color?: string;
  phase?: string;
  dueDate?: string | null;
  notes?: string;
  tags?: string[];
}

export interface UpdateProjectBody {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: ProjectStatus;
  color?: string;
  phase?: string;
  dueDate?: string | null;
  notes?: string;
  tags?: string[];
}

export interface Milestone {
  id: string;
  projectId: string;
  title: string;
  date: string;
  completed: boolean;
  createdAt: string;
}

export interface CreateMilestoneBody {
  /** @minLength 1 */
  projectId: string;
  /** @minLength 1 */
  title: string;
  date: string;
  completed?: boolean;
}

export interface UpdateMilestoneBody {
  /** @minLength 1 */
  title?: string;
  date?: string;
  completed?: boolean;
}

export interface SetMilestoneStatusBody {
  completed: boolean;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type EventVisibility =
  (typeof EventVisibility)[keyof typeof EventVisibility];

export const EventVisibility = {
  full: "full",
  shared: "shared",
} as const;

export interface Event {
  id: string;
  title: string;
  internalDescription?: string | null;
  sharedDescription: string;
  startDate: string;
  endDate: string;
  location: string;
  color: string;
  isAllDay: boolean;
  status: EventStatus;
  projectId?: string | null;
  createdByTeamId?: string | null;
  createdByUserId?: string | null;
  invitedTeamIds: string[];
  createdAt: string;
  updatedAt: string;
}

export type EventWithVisibility = Event & {
  visibility: EventVisibility;
};

export interface CreateEventBody {
  /** @minLength 1 */
  title: string;
  internalDescription?: string;
  sharedDescription?: string;
  startDate: string;
  endDate: string;
  location?: string;
  color?: string;
  isAllDay?: boolean;
  status?: EventStatus;
  projectId?: string | null;
  invitedTeamIds?: string[];
}

export interface UpdateEventBody {
  /** @minLength 1 */
  title?: string;
  internalDescription?: string;
  sharedDescription?: string;
  startDate?: string;
  endDate?: string;
  location?: string;
  color?: string;
  isAllDay?: boolean;
  status?: EventStatus;
  projectId?: string | null;
  invitedTeamIds?: string[];
}

export interface TeamNote {
  id: string;
  teamId: string;
  authorId: string;
  body: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTeamNoteBody {
  /** @minLength 1 */
  body: string;
}

export interface UpdateTeamNoteBody {
  /** @minLength 1 */
  body: string;
}

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Forbidden
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Conflict
 */
export type ConflictResponse = ErrorResponse;