# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ops & Planning API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectWithTeamName" }
    post:
      operationId: createProject
      tags: [projects]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectWithTeamName" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateProject
      tags: [projects]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteProject
      tags: [projects]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Milestones ───────────────────────────────────────────────────────
  /projects/{projectId}/milestones:
    get:
      operationId: listProjectMilestones
      tags: [milestones]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: projectId, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Milestone" }
  /milestones:
    post:
      operationId: createMilestone
      tags: [milestones]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMilestoneBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Milestone" }
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMilestoneBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Milestone" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /milestones/{id}/status:
    patch:
      operationId: setMilestoneStatus
      tags: [milestones]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetMilestoneStatusBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Milestone" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Events ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ProjectStatus:
      type: string
      enum: [not_started, in_progress, at_risk, completed]
    Project:
      type: object
      required: [id, teamId, title, description, status, color, phase, notes, tags, createdAt, updatedAt]
      properties:
        id: { type: string }
        teamId: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        color: { type: string }
        phase: { type: string }
        dueDate: { type: string, format: date-time, nullable: true }
        notes: { type: string }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectWithTeamName:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            teamName: { type: string, nullable: true }
    CreateProjectBody:
      type: object
      required: [teamId, title]
      properties:
        teamId: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        color: { type: string }
        phase: { type: string }
        dueDate: { type: string, format: date-time, nullable: true }
        notes: { type: string }
        tags:
          type: array
          items: { type: string }
    UpdateProjectBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        color: { type: string }
        phase: { type: string }
        dueDate: { type: string, format: date-time, nullable: true }
        notes: { type: string }
        tags:
          type: array
          items: { type: string }
    Milestone:
      type: object
      required: [id, projectId, title, date, completed, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        date: { type: string, format: date-time }
        completed: { type: boolean }
        createdAt: { type: string, format: date-time }
    CreateMilestoneBody:
      type: object
      required: [projectId, title, date]
      properties:
        projectId: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        date: { type: string, format: date-time }
        completed: { type: boolean }
    UpdateMilestoneBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        date: { type: string, format: date-time }
        completed: { type: boolean }
    SetMilestoneStatusBody:
      type: object
      required: [completed]
      properties:
        completed: { type: boolean }
...