# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ops & Planning API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from "./projectStatus";

export interface UpdateProjectBody {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: ProjectStatus;
  color?: string;
  phase?: string;
  dueDate?: Date | null;
  notes?: string;
  tags?: string[];
}